Reorganise data structures
Data structures have been moved and files renamed to be more consistent. Data
are separated between data structures related to a block and to the rescue in
general. Position, Size, Status and Operation are classes instead of enums and
basic types. This provide type safety for operations such as Position + Size
-> Size.