Optimize paintEvent to be usable
All square colors are computed first and squares are painted only after. With
small window sizes containing a few hundred squares, the widget can be painted
when resizing the window. With 3000 squares, it takes about one second to paint
the widget. With a full screen window containing 15000 squares, it takes about
7 seconds. It is usable but not smooth.