Display a QTableView instead of the grid view
The grid view will be a standard QTableView with a custom delegate to display
the colored grid (like in the Pixelator example). The grid will use the table
grid. The number of columns and rows returned by the model will depend on the
window size. There is no need to store rescue map data in QStandardItems any
more.