Remove RescueMapView border to get exact width
The RescueMapView now blends on the window and squares are shown entirely on
the right and bottom edge when the window is resized (it was not always the
case because the width and height included the QTableView border). The squares
can still be partially hidden behind the scrollbar when the square size is
modified.

RescueMapView background should be changed to have the same color than the
window border. Currently the white background can sometimes be seen on the
right and bottom edge of the grid while the window border is usually grey.