Optimize square color computation
Colors are computed faster by looping the rescue area only once. The colors
are also saved. With 4 times more squares, painting takes about half the time
than before the change to QTableView (about 4 seconds for 90000 squares
compared to about 8 seconds for 45000 squares on my computer). This is still
a bit slow to be used with 4-pixel squares and full-screen window.