Use uniform for tex_size
Pass the texture size tex_size as a uniform to the fragment shader.
It works as long as 3 <= tex_size <= 256.

It works with larger texture sizes such as 512 but the texture
components need to be changed to 'u2' (else the colors appear darker)
and the graphics card needs to be powerful enough because OpenGL 3.3
only requires a minimum size of 256x256x256 for 3D textures (see:
https://www.khronos.org/registry/OpenGL/specs/gl/glspec33.core.pdf#page=329).
Creating and loading a 512x512x512 texture may take a long time
(~30 seconds on my graphic card compared to ~3 seconds for 256 size).