Mini-map draft
Draw a mini-map with a scrolling bar on the left side. Currently it
displays a red scrolling rectangle to represent the current screen
position on the canvas. The rescue domain is presented vertically
on the scrolling area but the actual statuses are not correctly
displayed. Unused zoom uniforms have been removed. Mouse wheel does
not yet change the actual zoom because doing so crashes the app.