Working draft for the scrolling mini-map
Working draft for the scrolling mini-map. This works well with the
example.log which has many different statuses. When there are very
few sectors with specific statuses such as in Seagate1.mapfile,
the lines with those statuses may not be sampled from the texture.
As a result, there may not be color lines on the mini-map. Each
mini-map 1-pixel line should sample and OR the statuses for all
the texture pixels which represent this mini-map line.

Explaination for the rescue domain dark pixel has been found: the
status color for untried sectors -- vec4(0.25, 0.25, 0.25, 1.0) --
appears a lot darker with OpenGl (almost black) than the same color
in the original ddrescueview (medium grey).